Moved base classes to seperate folder, added include directory to CMakeLists.txt.
