+++ LogicCode/include/base_classes/object_representation_base.h
