+++ LogicCode/include/derivative.h
@@ -2,7 +2,7 @@
 // Description : Derivative class.
 
 #pragma once
-#include "sim_object_base.h"
+#include "base_classes/sim_object_base.h"
 
 class Derivative : public SimumlationObjectBase
 {
