+++ LogicCode/include/integrator.h
@@ -2,7 +2,7 @@
 // Description : Integrator class.
 
 #pragma once
-#include "sim_object_base.h"
+#include "base_classes/sim_object_base.h"
 
 class Integrator : public SimumlationObjectBase
 {
