+++ LogicCode/include/two_position_controller.h
@@ -2,7 +2,7 @@
 // Description : Two-position controller class.
 
 #pragma once
-#include "controller_base.h"
+#include "base_classes/controller_base.h"
 
 struct Two_Position_Controller_Params
 {
