+++ Tests/test_suites/control_system_with_two_position_controller_test.cpp
@@ -3,9 +3,9 @@
 
 #include <gtest/gtest.h>
 #include "test_with_plot.h"
-#include "include/signals.h"
-#include "include/two_position_controller.h"
-#include "include/object_differential_equation_representation.h"
+#include "signals.h"
+#include "two_position_controller.h"
+#include "object_differential_equation_representation.h"
 
 static constexpr uint8_t order = 2u;
 class ControlSystemWithTwoPositionControllerTest :  public testing::Test, public TestWithPlot
